long to int

+++ src/IdentityModel/EpochTimeExtensions.cs
@@ -13,29 +13,29 @@ namespace IdentityModel
         /// <summary>
         /// Converts the given date value to epoch time.
         /// </summary>
-        public static long ToEpochTime(this DateTime dateTime)
+        public static int ToEpochTime(this DateTime dateTime)
         {
             var date = dateTime.ToUniversalTime();
             var ticks = date.Ticks - new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc).Ticks;
             var ts = ticks / TimeSpan.TicksPerSecond;
-            return ts;
+            return (int)ts;
         }
 
         /// <summary>
         /// Converts the given date value to epoch time.
         /// </summary>
-        public static long ToEpochTime(this DateTimeOffset dateTime)
+        public static int ToEpochTime(this DateTimeOffset dateTime)
         {
             var date = dateTime.ToUniversalTime();
             var ticks = date.Ticks - new DateTimeOffset(1970, 1, 1, 0, 0, 0, TimeSpan.Zero).Ticks;
             var ts = ticks / TimeSpan.TicksPerSecond;
-            return ts;
+            return (int)ts;
         }
 
         /// <summary>
         /// Converts the given epoch time to a <see cref="DateTime"/> with <see cref="DateTimeKind.Utc"/> kind.
         /// </summary>
-        public static DateTime ToDateTimeFromEpoch(this long intDate)
+        public static DateTime ToDateTimeFromEpoch(this int intDate)
         {
             var timeInTicks = intDate * TimeSpan.TicksPerSecond;
             return new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc).AddTicks(timeInTicks);
@@ -44,7 +44,7 @@ namespace IdentityModel
         /// <summary>
         /// Converts the given epoch time to a UTC <see cref="DateTimeOffset"/>.
         /// </summary>
-        public static DateTimeOffset ToDateTimeOffsetFromEpoch(this long intDate)
+        public static DateTimeOffset ToDateTimeOffsetFromEpoch(this int intDate)
         {
             var timeInTicks = intDate * TimeSpan.TicksPerSecond;
             return new DateTimeOffset(1970, 1, 1, 0, 0, 0, TimeSpan.Zero).AddTicks(timeInTicks);
